refactor: hoist the plural-agreement helper into the core report module
